PBSJob: added job_main function.
SerialJobFactory: job checks if it was conducted on the same host on which it was created and if so uses cp instead of scp to copy results back.

Added blcr_checkpoint_script and blcr_restart_script scripts.

Added degen_primer_run -- serial job launcher for degen_primer.

setup.py: changed version to 1.1